Create slim docker container for go-ssb-room
This is a first iteration of a containerized app that can be pushed
to a docker repo and then referenced in a docker-compose.  I built
off the existing Dockerfile, but did it as a multi-stage build so that
the resulting image was smaller (300mb->30mb).

The existing start script assumes a relative path from the home directory
and an .env file.  This env file is unneccessary with the docker
container,  as we can just pass in the env vars when we run it.  We also
don't want our image  assume relative paths or copied in files.  I moved the
existing start script to start-dev.sh, then created a new start.sh that
better fits the container model.